+++ History.md
@@ -1,3 +1,7 @@
+0.1.0 - March 21, 2014
+-------------------------
+- adding first and second place winners
+- renaming oldest ticket
 
 0.0.1 - February 27, 2014
 -------------------------